requirements.txt
# Core
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0

# NLP & Embeddings
sentence-transformers>=2.6.0
scikit-learn>=1.4.0
spacy>=3.7.0

# Document parsing
PyMuPDF>=1.23.0          # PDF extraction (fitz)
python-docx>=1.1.0       # DOCX parsing

# Visualization
plotly>=5.20.0

# Notebook support
jupyter>=1.0.0
nbconvert>=7.0.0
ipykernel>=6.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
